Update Retrofit api and model

+++ 04-baking-app/app/src/main/java/fyi/jackson/drew/rezept/model/Ingredient.java
@@ -2,17 +2,17 @@ package fyi.jackson.drew.rezept.model;
 
 class Ingredient {
 
-    int quantity;
+    double quantity;
     String measure;
     String ingredient;
 
     public Ingredient() {}
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
